{{cookiecutter.project_name}}/requirements.txt
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
billiard==4.2.0
casbin==1.33.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
ecdsa==0.18.0
fastapi==0.104.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
kombu==5.3.3
limits==3.6.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
password-library==0.3.0
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary==3.1.12
puremagic==1.15
pyasn1==0.5.0
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9
simpleeval==0.9.13
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ulid==1.1
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.9
wrapt==1.15.0
